+++ components/dashboard/new-asset/forms/symbol-search-form.tsx
@@ -21,7 +21,7 @@ import { DialogBody, DialogFooter } from "@/components/ui/custom/dialog";
 import { YahooFinanceLogo } from "@/components/ui/logos/yahoo-finance-logo";
 import { LocalizedNumberInput } from "@/components/ui/custom/localized-number-input";
 import { SymbolSearch } from "../../symbol-search";
-import { PositionCategorySelector } from "@/components/dashboard/position-category-selector";
+import { PositionCategorySelector } from "@/components/dashboard/categories/position-category-selector";
 import { CapitalGainsTaxRateField } from "@/components/dashboard/positions/shared/capital-gains-tax-rate-field";
 
 import { useNewAssetDialog } from "../index";
